// src/app/model/movie-details.ts
export interface MovieDetails {
  backdrop_path: string;
  budget: number;
  homepage: string;
  id: number;
  imdb_id: string;
  original_lan: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: any;
  production_countries: any;
  release_date: string;
  revenue: number;
  runtime: number;
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  genre: any;
  background_image: any;
  watchprovider: any;
  reviewList: any;

  backdropList: any;
  logoList: any;
  posterList: any;
  castList: any;
  crewList: any;
  similarmovieList: any;
  recmovieList: any;
  videoList: any;
}